update (and simplify) dependenies
