+++ mmda-backend/Pipfile
@@ -4,37 +4,33 @@ url = "https://pypi.python.org/simple"
 verify_ssl = true
 
 [dev-packages]
-pytest = "==7.0.1"
-pylint = "==2.13.9"
-pytest-cov = "==3.0.0"
-sphinx = "==5.0.0"
+pytest = "==7.1.3"
+pylint = "==2.15.4"
+pytest-cov = "==4.0.0"
+sphinx = "==5.2.3"
 
 [packages]
-Flask = "==1.1.2"
-Flask-Migrate = "==2.3.1"
-Flask-Script = "==2.0.6"
-Flask-SQLAlchemy = "==2.5.1"
-Flask-WTF = "==0.14.3"
-Flask-User = "==1.0.2.2"
-Flask-JWT-Extended = "==4.1.0"
-Flask-Caching = "==1.10.1"
-Flask-Cors = "==3.0.7"
-flask-expects-json = "==1.5.0"
-jsonschema = "==3.2.0"
-email_validator = "==1.1.1"
-jinja2 = "==3.0.1"
-itsdangerous = "==2.0.1"
-werkzeug = "==0.16.1"
-gevent = "==21.1.2"
-pathlib2 = "==2.3.5"
-pandas = "==1.2.0"
+Flask = "==2.2.2"
+Flask-Login = "==0.6.2"
+Flask-SQLAlchemy = "==3.0.1"
+Flask-WTF = "==1.0.1"
+Flask-JWT-Extended = "==4.4.4"
+Flask-Mail = "==0.9.1"
+Flask-Cors = "==3.0.10"
+flask-expects-json = "==1.7.0"
+
+pandas = "==1.5.0"
 pymagnitude-lite = "==0.1.143"
-scikit-learn = "==0.23.1"
-umap-learn = "==0.5.1"
+scikit-learn = "==1.1.2"
+umap-learn = "==0.5.3"
 anycache = "==2.0.7"
+
+gevent = "==22.8.0"
 gunicorn = "==20.1.0"
-cwb-ccc = {git = "https://github.com/ausgerechnet/cwb-ccc.git", ref = "v0.10.3"}
-# cwb-ccc = {editable = true, path = "./../../cwb-ccc"}
+
+# cwb-ccc = "==0.10.3"
+cwb-ccc = {editable = true, path = "./../../cwb-ccc"}
+association-measures = "==0.2.5"
 
 [requires]
-python_version = "3.8"
+python_version = "3.9"
