--- mmda-backend/tests/analysis/collocates-atomkraft-germaparl1318.txt
@@ -1,524 +1,0 @@
-17
-32
-50
-435
-436
-437
-440
-442
-444
-2011
-2020
-2022
-&quot;
-&
-,
--Reduktion
--frei
--freie
--freien
--neutral
-.
-000-Dächer-Programm
-AKW
-Abfall
-Abklingbecken
-Abschalten
-Abschaltplan
-Abschaltung
-Alterung
-Angra
-Anteil
-Arbeitsplatzverlust
-Aspekt
-AtomMüll
-Atomausstieg
-Atomgesetz
-Atomkonsens
-Atommüllproblem
-Atomstrom
-Atomwirtschaft
-Auffassung
-Ausbau
-Ausbreitung
-Auslaufbetrieb
-Auslaufmodell
-Ausstieg
-Barsebäck
-Bau
-Bauart
-Beendigung
-Befürworter
-Befürwortung
-Beginn
-Beschluss
-Betreiben
-Betreiber
-Betreibergesellschaft
-Betrieb
-Betriebsgenehmigung
-Biblis
-Block
-Brasilien
-Braunkohle
-Brennelement
-Brennstoff
-Brokdorf
-Brunsbüttel
-Brücke
-Brückentechnologie
-CO2
-Castor-Transport
-Castortransport
-Cattenom
-Co.
-Deutschland
-EVUs
-Einstieg
-Elektrizität
-Emission
-Ende
-Endlager
-Endlagerkapazität
-Endlagerung
-Energie
-Energieerzeugung
-Energieform
-Energieimport
-Energiekonsens
-Energiekonzept
-Energiekonzern
-Energiemix
-Energiepolitik
-Energiestruktur
-Energieträger
-Energieversorgung
-Energieversorgungsunternehmen
-Energiewende
-Entscheidung
-Entsorgung
-Entsorgungsfrage
-Entsorgungsproblem
-Erdbebengebiet
-Erneuerbar
-Ersatz
-Erzeugung
-Euratom
-Fessenheim
-Festhalt
-Finnland
-Flugzeugabstürz
-Folge
-Folgelast
-Forsmark
-Frankreich
-Fukushima
-Förderung
-Gas
-Gefahr
-Gegner
-Gesprächsgegenstand
-GmbH
-Grafenrheinfeld
-Greifswald
-Grohnde
-Grund
-Grundlast
-Gründungsthema
-Gundremmingen
-Hermes-Bürgschaft
-HermesBürgschaft
-Hinkley
-Hochrisikotechnologie
-Inbetriebnahme
-Informationskreis
-Isar
-Jahr
-Jahrzehnt
-Japan
-K
-Kampfthema
-Kernforschung
-Kernfusion
-Khmelnitski
-Khmelnytsky
-Kilowattstunde
-Klimaschutz
-Klimaschutzziel
-Know-how
-Koalitionsvertrag
-Kohle
-Kohle-
-Kohleenergie
-Kohlekraft
-Kohlekraftwerk
-Kohleverstromung
-Konsens
-Konzern
-Kozloduj
-Kraftwerk
-Krümmel
-Lagerung
-Laufzeit
-Laufzeitbegrenzung
-Laufzeitverlängerung
-Leistungsbetrieb
-Lingen
-Lubmin
-Lösungsansatz
-Mehrheit
-Mochovce
-Mülheim-Kärlich
-Müll
-Nachbarstaat
-Nachrüstung
-Neckarwestheim
-Netz
-Netzausbau
-Neubau
-Neubewertung
-Normalbetrieb
-Nuklearanlage
-Nutzung
-Nähe
-OHG
-Obrigheim
-Option
-Osteuropa
-Philippsburg
-Point
-Raus
-Reaktor
-Renaissance
-Restbetrieb
-Restlaufzeit
-Restrisiko
-Reststrommenge
-Rheinsberg
-Risiko
-Risikotechnologie
-Rot-Grün
-Rowno
-Rückbau
-Rückstellung
-Sache
-Schweden
-Sicherheit
-Sicherheitsanforderung
-Sicherheitsbedenken
-Sicherheitsgesichtspunkt
-Sicherheitskonzeption
-Sicherheitskultur
-Sicherheitsmanagementsystem
-Sicherheitsniveau
-Sicherheitsstandard
-Sicherheitstechnik
-Sicherheitsüberprüfung
-Stade
-Stillegung
-Stilllegung
-Stresstest
-Strom
-Strombrücke
-Stromerzeugung
-Stromerzeugungsenergie
-Strommenge
-Stromproduktion
-Stromversorgung
-Stufenplan
-Störfall
-Subvention
-Szenario
-Süden
-Technologie
-Temelin
-Thema
-Tihange
-Transport
-Tschernobyl
-Ukraine
-Umbau
-Umgebung
-Umstieg
-Unfall
-Unterweser
-Unverzichtbarkeit
-Uran
-Verankerung
-Verlängerung
-Verminderung
-Versorgungssicherheit
-Weiterbetrieb
-Weiterlaufen
-Welt
-Wenn
-Wiederaufarbeitung
-Wiederaufbereitung
-Wiedereinstieg
-Willkür
-Würgassen
-Yuzhno-Ukrainska
-Zeitalter
-Zwischenlager
-Zwischenlagerung
-ab
-abbauen
-abgebrannt
-abgeschaltet
-abgeschrieben
-abschalten
-abstellen
-acht
-als
-alt
-anfallend
-angestrebt
-anzweifeln
-auf
-aus
-ausgehen
-auskommen
-auslaufen
-ausschalten
-aussprechen
-aussteigen
-bauen
-baugleich
-bedeuten
-beenden
-befindlich
-befrieden
-befürworten
-beherrschbar
-beilegen
-beinhalten
-bejahen
-belgisch
-bescheren
-beschleunigen
-beschleunigt
-beschließen
-beschlossen
-besprechen
-bestehend
-bestellen
-bestellt
-bestrahlt
-betreffend
-betreiben
-betreibend
-betrieben
-billig
-bis
-brasilianisch
-britisch
-bzw.
-chinesisch
-dass
-decken
-defekt
-demonstrieren
-deutsch
-die
-durch
-durchgesetzt
-durchsetzen
-ein
-eine
-einhergehen
-einleiten
-einsetzen
-einsteigen
-eintreten
-einzig
-endgültig
-energetisch
-entscheiden
-entschädigungsfrei
-erfolgen
-erlöschen
-erneuerbar
-ersetzen
-erzeugen
-erzeugt
-erzwingen
-festhalten
-final
-finnisch
-fordern
-fossil
-französisch
-friedlich
-früh
-funktionsfähig
-für
-geben
-gefährlich
-gegen
-gehören
-geliebt
-generell
-geordnet
-geplant
-gewerblich
-gewollt
-glaubhaft
-gleichzeitig
-grenznah
-grundgesetzlich
-grundlastfähig
-halten
-heraus
-herauskommen
-herum
-hierzulande
-hinsichtlich
-hochfahren
-ideologisch
-in
-insgesamt
-investieren
-japanisch
-keine
-klimafreundlich
-klimaneutral
-klimapolitisch
-klimaverträglich
-kommerziell
-kompensieren
-konsensual
-kontra
-kontrolliert
-lang
-langfristig
-laufen
-laufend
-leisten
-liegend
-marode
-meiste
-militärisch
-mit
-mittelfristig
-möglich
-möglichst
-neu
-nutzen
-nämlich
-obwohl
-oder
-offenhalten
-ohne
-organisieren
-ort
-osteuropäisch
-preiswert
-produzieren
-propagieren
-protestieren
-radioaktiv
-rational
-raus
-regenerativ
-riskant
-rückgängig
-schnell
-schnellstmöglich
-schrittweise
-schwedisch
-sein
-setzen
-sicher
-sicherheitstechnisch
-sieben
-slowakisch
-sofort
-sofortig
-sowjetisch
-sozialistisch
-stammen
-statt
-stehen
-stillgelegt
-stilllegen
-streiten
-störanfällig
-sukzessiv
-sämtliche
-technisch
-teilweis
-terroranfällig
-ukrainisch
-umgehen
-umweltfreundlich
-unbefristet
-unbegrenzt
-unbeherrschbar
-und
-uneingeschränkt
-ungelöst
-unsicher
-unsinnig
-unterschiedlich
-untrennbar
-unumkehrbar
-unver
-unverantwortlich
-unverzüglich
-unwirtschaftlich
-usw.
-verabschieden
-veraltet
-verantwortbar
-verbinden
-vereinbaren
-vereinbart
-verfrüht
-verknüpfen
-verkürzen
-verlängern
-verlängert
-verschwenderisch
-verstopfen
-vertretbar
-verursachen
-verzichten
-vollziehen
-von
-vorantreiben
-vorbereiten
-vorhanden
-vorzeitig
-weit
-weiter
-weiterbetreiben
-weiterhin
-weiterlaufen
-weltweit
-wenn
-werden
-werk
-werken
-wie
-willkürlich
-wohnen
-wollen
-während
-zehn
-zeitgleich
-zementieren
-zivil
-zu
-zurückbauen
-zurückgebaut
-zurückkehren
-zusätzlich
-zwingen
-zwölf
-zügig
-Ökoenergie
-Überbrückungszeit
-Übergangsenergie
-über
-überdenken
-übergeordnet
