+++ mmda-backend/tests/conftest.py
@@ -1,14 +1,19 @@
-# This file contains pytest 'fixtures'.
-# If a test functions specifies the name of a fixture function as a parameter,
-# the fixture function is called and its result is passed to the test function.
-
-import pytest
-from backend import create_app
-from backend.commands.init_db import init_db
 import os
 
+import pytest
+from flask_jwt_extended import create_access_token
+
+from backend import create_app
+from backend.models.database import db as the_db
+from backend.models.database import init_db
 
 here = os.path.dirname(os.path.realpath(__file__))
+
+
+# Init test vectors
+test_vectors = open('/tmp/foo.pymagnitude', 'w')
+test_vectors.write('nothing to see here')
+test_vectors.close()
 
 
 the_app = create_app(dict(
@@ -22,10 +27,13 @@ the_app = create_app(dict(
     WTF_CSRF_ENABLED=False,  # Disable CSRF form validation
 ))
 
+
 # Setup an application context (since the tests run outside of the webserver context)
 the_app.app_context().push()
 
 # Create and populate roles and users tables
+init_db()
+# Run twice to ensure the function runs idempotently
 init_db()
 
 
@@ -33,6 +41,59 @@ init_db()
 def app():
     """ Makes the 'app' parameter available to test functions. """
     return the_app
+
+
+@pytest.fixture(scope='session')
+def db():
+    """ Makes the 'db' parameter available to test functions. """
+    return the_db
+
+
+@pytest.fixture(scope='function')
+def session(db, request):
+    """Creates a new database session for a test."""
+    connection = db.engine.connect()
+    transaction = connection.begin()
+
+    options = dict(bind=connection, binds={})
+    session = db.create_scoped_session(options=options)
+
+    db.session = session
+
+    def teardown():
+        transaction.rollback()
+        connection.close()
+        session.remove()
+
+    request.addfinalizer(teardown)
+    return session
+
+
+@pytest.fixture(scope='session')
+def client(app):
+    return app.test_client()
+
+
+@pytest.fixture
+def header():
+    access_token = create_access_token('student1')
+
+    headers = {
+        'Authorization': 'Bearer {}'.format(access_token)
+    }
+
+    return headers
+
+
+@pytest.fixture
+def admin_header():
+    access_token = create_access_token(identity={'username': 'admin', 'roles': ['admin']})
+
+    headers = {
+        'Authorization': 'Bearer {}'.format(access_token)
+    }
+
+    return headers
 
 
 @pytest.fixture
@@ -68,7 +129,8 @@ def test_corpus():
         'Anpassung': ["Anpassungsmaßnahme", "Maßnahme", "bewältigen", "Anpassung"]
     }
 
-    collocates_atomkraft = set(open(os.path.join(here, "collocates-atomkraft-germaparl1318.txt")).read().split("\n"))
+    with open(os.path.join(here, "corpora", "collocates-atomkraft-germaparl1318.txt")) as f:
+        collocates_atomkraft = set(f.read().split("\n"))
 
     return {
         'corpus_name': corpus_name,
