+++ mmda-backend/tests/test_views_keyword.py
@@ -1,5 +1,5 @@
-from flask import url_for
 import pytest
+from flask import url_for
 
 
 @pytest.mark.api
